+++ oap-server/server-core/src/main/java/org/apache/skywalking/oap/server/core/storage/StorageModule.java
@@ -19,6 +19,7 @@
 package org.apache.skywalking.oap.server.core.storage;
 
 import org.apache.skywalking.oap.server.core.storage.cache.INetworkAddressAliasDAO;
+import org.apache.skywalking.oap.server.core.storage.management.UIMenuManagementDAO;
 import org.apache.skywalking.oap.server.core.storage.management.UITemplateManagementDAO;
 import org.apache.skywalking.oap.server.core.storage.profiling.continuous.IContinuousProfilingPolicyDAO;
 import org.apache.skywalking.oap.server.core.storage.profiling.ebpf.IServiceLabelDAO;
@@ -75,6 +76,7 @@ public class StorageModule extends ModuleDefine {
             IProfileTaskLogQueryDAO.class,
             IProfileThreadSnapshotQueryDAO.class,
             UITemplateManagementDAO.class,
+            UIMenuManagementDAO.class,
             IBrowserLogQueryDAO.class,
             IEventQueryDAO.class,
             IEBPFProfilingTaskDAO.class,